Refactor astrometry backends
