+++ skylib/astrometry/atlas/__init__.py
@@ -1,3 +1,4 @@
+from .backend import AtlasBackend
 from .config import AtlasConfig
 
-__all__ = ["AtlasConfig"]
+__all__ = ["AtlasBackend", "AtlasConfig"]
